:truck: feat: refactor property manipulation into new class for reuse
